refactor(converters): extract surface partitioner and decompose logic
- Extract surface partitioning and Canvas extraction from a2a-ui-converter
  into a dedicated surface-partitioner module, re-exporting for callers.
- Decompose partitionA2uiSurfacePayload into smaller focused helper
  functions (categorizeSurfaceItems, buildComponentMap, collectDescendantIds,
  formatCanvasComponents, extractCanvasArtifacts, buildInlineSurfacePayload).
